updateAllCHanges

+++ src/app/superadmin/pages/supquestion-list/supquestion-list.component.ts
@@ -5,35 +5,40 @@ import { AddQuestionComponent } from '../../project/Components/add-question/add-
 import { ToastrService } from 'ngx-toastr';
 import { ApiService } from '../../services/api.service';
 import { SearchService } from '../../services/search.service';
+import { DeleteComponent } from '../delete/delete.component';
 
 @Component({
   selector: 'app-supquestion-list',
   templateUrl: './supquestion-list.component.html',
-  styleUrl: './supquestion-list.component.css'
+  styleUrl: './supquestion-list.component.css',
 })
 export class SupquestionListComponent {
-  data:any;
-  mcqdescriptive:any;
-  descriptive:any;
-  mcq:any;
-  page:number=1;
-  isLoading:boolean=false;
-  constructor(private api:ProjectService,private dialog:MatDialog,private tosatr:ToastrService,private service:ApiService,private servicesearch:SearchService){}
-  
+  data: any;
+  mcqdescriptive: any;
+  descriptive: any;
+  mcq: any;
+  page: number = 1;
+  isLoading: boolean = false;
+  constructor(
+    private api: ProjectService,
+    private dialog: MatDialog,
+    private tosatr: ToastrService,
+    private service: ApiService,
+    private servicesearch: SearchService
+  ) {}
+
   ngOnInit(): void {
-    this.isLoading=true
-    this.api.getAllQuestions().subscribe((res)=>{
-      if(res.success){
-        this.isLoading=false
-        this.data=res.data;
+    this.isLoading = true;
+    this.api.getAllQuestions().subscribe((res) => {
+      if (res.success) {
+        this.isLoading = false;
+        this.data = res.data;
         console.log(this.data);
       }
-    })
-    
+    });
+
     this.service.getCountQuestions().subscribe((res: any) => {
-  
       if (res.success) {
-       
         this.mcq = res.data.mcq;
         this.descriptive = res.data.descriptive;
         this.mcqdescriptive = res.data.mcqAndDescriptive;
@@ -41,15 +46,14 @@ export class SupquestionListComponent {
       }
     });
 
-
     this.servicesearch.sendResults().subscribe({
       next: (res: any) => {
         if (res.length == 0) {
-          this.isLoading=false
-          this.getAllQues()
+          this.isLoading = false;
+          this.getAllQues();
         } else {
           if (res.success) {
-            this.isLoading=false
+            this.isLoading = false;
             this.data = res.data;
           } else {
             this.data = [];
@@ -59,55 +63,65 @@ export class SupquestionListComponent {
       error: (err: any) => {},
       complete: () => {},
     });
-
   }
 
-  getAllQues(){
-    this.api.getAllQuestions().subscribe((res:any)=>{
-      if(res.success){
-        this.data=res.data;
+  getAllQues() {
+    this.api.getAllQuestions().subscribe((res: any) => {
+      if (res.success) {
+        this.data = res.data;
         console.log(this.data);
-        
       }
-    })
-  }
-  deleteQuestion(id:any){
-this.api.deleteQuestion(id).subscribe((res:any)=>{
-  console.log(res);
-  if(res.message==="Question deleted successfully."){
-    this.tosatr.success(res.message);
-    this.getAllQues();
-    // window.location.reload();
-  }
-  
-})
+    });
   }
 
-  addQuestion(){
+  deleteQuestion(id: any) {
+    const dialogRef = this.dialog.open(DeleteComponent, {
+      data: {
+        message: `Do you really want to delete the record ?`,
+      },
+      disableClose:true
+    });
+  
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result.action == 'ok') {
+        this.api.deleteQuestion(id).subscribe((res: any) => {
+          console.log(res);
+          if (res.message === 'Question deleted successfully.') {
+            this.tosatr.success(res.message);
+            this.getAllQues();
+            // window.location.reload();
+          }
+        });
+      }
+    });
+  }
+  addQuestion() {
     const dialogRef = this.dialog.open(AddQuestionComponent, {
       width: '1100px',
       height: '700px',
       disableClose: true,
     });
     dialogRef.afterClosed().subscribe(() => {
-     this.getAllQues();
+      this.getAllQues();
     });
   }
-  
-  getQuestionByStatus(status:any){
-    this.api.getQuestionListByStatus(status).subscribe((res:any)=>{
-      this.data=null;
-      if(res.success){
-        this.data=res.data;
-        console.log('Client by status=>'+res.data)
-        console.log(res.message);
+
+  getQuestionByStatus(status: any) {
+    this.api.getQuestionListByStatus(status).subscribe(
+      (res: any) => {
+        this.data = null;
+        if (res.success) {
+          this.data = res.data;
+          console.log('Client by status=>' + res.data);
+          console.log(res.message);
+        } else {
+          this.tosatr.error(res.message);
+        }
+      },
+      (error) => {
+        this.tosatr.error('Clients Not Found..!!');
       }
-      else{
-        this.tosatr.error(res.message);
-      }
-    },(error)=>{
-      this.tosatr.error('Clients Not Found..!!');
-    })
+    );
   }
 
   pageChangeEvent(event: number) {
